requirements.txt
# Core App & UI
streamlit==1.47.0

# LangChain & Google Gemini
langchain==0.3.26
langchain-core==0.3.69
langchain-community==0.3.27
langchain-google-genai==2.1.8
google-generativeai==0.4.1
google-api-core==2.25.1
google-auth==2.40.3

# Pinecone Vector DB for RAG
pinecone==7.3.0

# Document Parsing (PDF & DOCX)
PyMuPDF==1.26.3
pypdf==5.8.0
python-docx==1.2.0
docx2txt==0.9
python-magic==0.4.27

# Email Handling & Parsing
beautifulsoup4==4.13.4
html5lib==1.1
secure-smtplib==0.1.1

# Text & Utility
tiktoken==0.9.0
langdetect==1.0.9
regex==2024.11.6
tenacity==9.1.2
python-dotenv==1.1.1
typing_extensions==4.14.1
